requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
asyncio==3.4.3
Django==5.0
frozenlist==1.4.1
fuzzywuzzy==0.18.0
logging==0.4.9.6
matplotlib==3.8.3
pillow==10.2.0
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.1
python-Levenshtein==0.23.0
pytz==2024.1
paramiko